+++ CHANGELOG.md
@@ -1,8 +1,13 @@
-# unreleased
+# unreleased 
+
+## Features
+
+* New `--print-color-table` option, see #229 (@sahinfalcon)
 
 ## Bugfixes
 
 - Throw an error when try to view a directory, see #234 (@Integral-Tech)
+
 
 # v0.15.0